Lint ImageWidget, move ImageType enum into its own file
