+++ app/src/org/odk/collect/android/activities/FormEntryActivity.java
@@ -1770,7 +1770,7 @@ public class FormEntryActivity extends CommCareActivity<FormEntryActivity>
         }
     }
 
-    public void handleFormLoadCompletion(FormController fc) {
+    private void handleFormLoadCompletion(FormController fc) {
         mFormController = fc;
         if(Build.VERSION.SDK_INT >= Build.VERSION_CODES.HONEYCOMB){
             // Newer menus may have already built the menu, before all data was ready
